requirements.txt
asgiref==3.6.0
Django==4.2.1
django-dotenv==1.4.2
djangorestframework==3.14.0
install==1.3.5
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
